Update unit & e2e tests
