+++ src/applications/appeals/995/tests/995-supplemental-claim.cypress.spec.js
@@ -176,10 +176,7 @@ const testConfig = createTestConfig(
                 }
               }
             });
-            cy.get('va-button-pair')
-              .shadow()
-              .find('va-button[continue]')
-              .click();
+            cy.findByText('Continue', { selector: 'button' }).click();
           });
         });
       },
