+++ src/applications/appeals/995/tests/components/EvidencePrivateRecords.unit.spec.jsx
@@ -17,6 +17,7 @@ import {
 
 import { getDate } from '../../../shared/utils/dates';
 import { SELECTED } from '../../../shared/constants';
+import { clickAddAnother, clickBack, clickContinue } from './helpers';
 
 /*
 | Data     | Forward     | Back               | Add another      |
@@ -200,7 +201,7 @@ describe('<EvidencePrivateRecords>', () => {
       const { container } = render(page);
 
       // continue
-      fireEvent.click($('.usa-button-primary', container));
+      clickContinue(container);
       await waitFor(() => expect(goSpy.calledWith(data)).to.be.true);
     });
     it('should navigate back to private records request page with valid data', async () => {
@@ -215,7 +216,7 @@ describe('<EvidencePrivateRecords>', () => {
       const { container } = render(page);
 
       // back
-      fireEvent.click($('.usa-button-secondary', container));
+      clickBack(container);
       // passing a negative index is okay, we're leaving the indexed pages
       await waitFor(() => expect(goSpy.calledWith(index - 1)).to.be.true);
     });
@@ -234,7 +235,7 @@ describe('<EvidencePrivateRecords>', () => {
       const { container } = render(page);
 
       // add
-      fireEvent.click($('.vads-c-action-link--green', container));
+      clickAddAnother(container);
 
       await waitFor(() => {
         expect($('va-modal[visible="false"]', container)).to.exist;
@@ -255,7 +256,7 @@ describe('<EvidencePrivateRecords>', () => {
       const { container } = render(page);
 
       // add
-      fireEvent.click($('.vads-c-action-link--green', container));
+      clickAddAnother(container);
 
       await waitFor(() => {
         expect($('va-modal[visible="false"]', container)).to.exist;
@@ -303,7 +304,7 @@ describe('<EvidencePrivateRecords>', () => {
       const { container } = render(page);
 
       // continue
-      fireEvent.click($('.usa-button-primary', container));
+      clickContinue(container);
 
       await waitFor(() => {
         expect($('va-modal[visible="false"]', container)).to.exist;
@@ -327,7 +328,7 @@ describe('<EvidencePrivateRecords>', () => {
       const { container } = render(page);
 
       // continue
-      fireEvent.click($('.usa-button-primary', container));
+      clickContinue(container);
 
       await waitFor(() => {
         expect($('va-modal[visible="false"]', container)).to.exist;
@@ -342,7 +343,7 @@ describe('<EvidencePrivateRecords>', () => {
       const { container } = render(page);
 
       // back
-      fireEvent.click($('.usa-button-secondary', container));
+      clickBack(container);
 
       await waitFor(() => {
         expect($('va-modal[visible="false"]', container)).to.exist;
@@ -364,7 +365,7 @@ describe('<EvidencePrivateRecords>', () => {
       const { container } = render(page);
 
       // back
-      fireEvent.click($('.usa-button-secondary', container));
+      clickBack(container);
 
       await waitFor(() => {
         expect($('va-modal[visible="false"]', container)).to.exist;
@@ -389,7 +390,7 @@ describe('<EvidencePrivateRecords>', () => {
       const { container } = render(page);
 
       // add
-      fireEvent.click($('.vads-c-action-link--green', container));
+      clickAddAnother(container);
 
       await waitFor(() => {
         expect($('va-modal[visible="false"]', container)).to.exist;
@@ -404,7 +405,7 @@ describe('<EvidencePrivateRecords>', () => {
       const { container } = render(page);
 
       // back
-      await fireEvent.click($('.usa-button-secondary', container));
+      clickBack(container);
 
       const event = new CustomEvent('closeEvent');
       await $('va-modal', container).__events.closeEvent(event);
@@ -450,7 +451,7 @@ describe('<EvidencePrivateRecords>', () => {
       const { container } = render(page);
 
       // continue
-      fireEvent.click($('.usa-button-primary', container));
+      clickContinue(container);
 
       await waitFor(() => {
         expect(goSpy.called).to.be.false;
@@ -460,7 +461,6 @@ describe('<EvidencePrivateRecords>', () => {
     });
 
     // *** BACK ***
-    // consistently flaky test
     it('should show modal, select "No, remove this location", then navigate back to previous index', async () => {
       const goSpy = sinon.spy();
       const setDataSpy = sinon.spy();
@@ -483,7 +483,7 @@ describe('<EvidencePrivateRecords>', () => {
       const { container } = render(page);
 
       // back
-      fireEvent.click($('.usa-button-secondary', container));
+      clickBack(container);
       // Click no
       await testAndCloseModal(container, 8, 'secondaryButtonClick');
 
@@ -496,7 +496,6 @@ describe('<EvidencePrivateRecords>', () => {
       });
     });
 
-    // consistently flaky test
     it('should show modal, select "Yes", then navigate back to previous index', async () => {
       const goSpy = sinon.spy();
       const setDataSpy = sinon.spy();
@@ -517,7 +516,7 @@ describe('<EvidencePrivateRecords>', () => {
       const { container } = render(page);
 
       // back
-      fireEvent.click($('.usa-button-secondary', container));
+      clickBack(container);
       // Click yes to keep partial entry
       await testAndCloseModal(container, 8, 'primaryButtonClick');
 
@@ -544,7 +543,7 @@ describe('<EvidencePrivateRecords>', () => {
       const { container } = render(page);
 
       // add
-      fireEvent.click($('.vads-c-action-link--green', container));
+      clickAddAnother(container);
 
       await waitFor(() => {
         expect(goSpy.called).to.be.false;
